watertrigger: add null checks to SAVControl to allow usage without it

+++ Scripts/SaccAirVehicle/SAV_Extensions/SAV_WaterTrigger.cs
@@ -16,9 +16,9 @@ namespace SaccFlightAndVehicles
         public float WaterSlowDown = 3;
         [Tooltip("Strength of force slowing down the vehicle's rotation when it's underwater")]
         public float WaterSlowDownRot = 3;
-        private SaccEntity EntityControl;
+        [System.NonSerialized] public SaccEntity EntityControl;
         private Rigidbody VehicleRigidbody;
-        private bool CFOverridden;
+        private bool ApplyPhysOverridden;
         private int WaterLayer = 0;
         private int NumTriggers = 0;
         private bool InWater;
@@ -29,7 +29,6 @@ namespace SaccFlightAndVehicles
         {
             Initilized = true;
             WaterLayer = LayerMask.NameToLayer("Water");
-            EntityControl = (SaccEntity)SAVControl.GetProgramVariable("EntityControl");
             VehicleRigidbody = EntityControl.GetComponent<Rigidbody>();
             ThisCollider = gameObject.GetComponent<Collider>();
             VRCPlayerApi localPlayer = Networking.LocalPlayer;
@@ -38,14 +37,13 @@ namespace SaccFlightAndVehicles
                 if (!localPlayer.isMaster)
                 { gameObject.SetActive(false); }
             }
-
         }
         private void Update()
         {
             if (InWater)
             {
                 float DeltaTime = Time.deltaTime;
-                SAVControl.SetProgramVariable("Health", (float)SAVControl.GetProgramVariable("Health") - (WaterDamageSec * DeltaTime));
+                if (SAVControl) { SAVControl.SetProgramVariable("Health", (float)SAVControl.GetProgramVariable("Health") - (WaterDamageSec * DeltaTime)); }
                 VehicleRigidbody.velocity = Vector3.Lerp(VehicleRigidbody.velocity, Vector3.zero, WaterSlowDown * DeltaTime);
                 VehicleRigidbody.angularVelocity = Vector3.Lerp(VehicleRigidbody.angularVelocity, Vector3.zero, WaterSlowDownRot * DeltaTime);
             }
@@ -58,10 +56,10 @@ namespace SaccFlightAndVehicles
                 if (!InWater)
                 { SendCustomNetworkEvent(VRC.Udon.Common.Interfaces.NetworkEventTarget.All, nameof(SendEnterWater)); }
                 InWater = true;
-                if (!CFOverridden)
+                if (!ApplyPhysOverridden)
                 {
-                    CFOverridden = true;
-                    SAVControl.SetProgramVariable("OverrideConstantForce", (int)SAVControl.GetProgramVariable("OverrideConstantForce") + 1);
+                    ApplyPhysOverridden = true;
+                    if (SAVControl) { SAVControl.SetProgramVariable("DisablePhysicsApplication", (int)SAVControl.GetProgramVariable("DisablePhysicsApplication") + 1); }
                 }
             }
         }
@@ -75,10 +73,10 @@ namespace SaccFlightAndVehicles
                     if (InWater)
                     { SendCustomNetworkEvent(VRC.Udon.Common.Interfaces.NetworkEventTarget.All, nameof(SendExitWater)); }
                     InWater = false;
-                    if (CFOverridden)
+                    if (ApplyPhysOverridden)
                     {
-                        CFOverridden = false;
-                        SAVControl.SetProgramVariable("OverrideConstantForce", (int)SAVControl.GetProgramVariable("OverrideConstantForce") - 1);
+                        ApplyPhysOverridden = false;
+                        if (SAVControl) { SAVControl.SetProgramVariable("DisablePhysicsApplication", (int)SAVControl.GetProgramVariable("DisablePhysicsApplication") - 1); }
                     }
                 }
             }
@@ -88,7 +86,7 @@ namespace SaccFlightAndVehicles
             EntityControl.SendEventToExtensions("SFEXT_G_EnterWater");
             if (!DisableTaxiRotation)
             {
-                SAVControl.SetProgramVariable("DisableTaxiRotation", (int)SAVControl.GetProgramVariable("DisableTaxiRotation") + 1);
+                if (SAVControl) { SAVControl.SetProgramVariable("DisableTaxiRotation", (int)SAVControl.GetProgramVariable("DisableTaxiRotation") + 1); }
                 DisableTaxiRotation = true;
             }
         }
@@ -97,7 +95,7 @@ namespace SaccFlightAndVehicles
             EntityControl.SendEventToExtensions("SFEXT_G_ExitWater");
             if (DisableTaxiRotation)
             {
-                SAVControl.SetProgramVariable("DisableTaxiRotation", (int)SAVControl.GetProgramVariable("DisableTaxiRotation") - 1);
+                if (SAVControl) { SAVControl.SetProgramVariable("DisableTaxiRotation", (int)SAVControl.GetProgramVariable("DisableTaxiRotation") - 1); }
                 DisableTaxiRotation = false;
             }
         }
@@ -110,25 +108,28 @@ namespace SaccFlightAndVehicles
         private void OnDisable()
         {
             ThisCollider.enabled = false;
-            if (InWater && WaterDamageSec > 0)
-            { SAVControl.SetProgramVariable("Health", -1); }//just kill the vehicle if it's underwater and the player gets out
+            if (SAVControl)
+            {
+                if (InWater && WaterDamageSec > 0)
+                { SAVControl.SetProgramVariable("Health", -1); }//just kill the vehicle if it's underwater and the player gets out
+            }
             InWater = false;
             NumTriggers = 0;
         }
         public void SFEXT_G_Explode()
         {
-            if (CFOverridden)
+            if (ApplyPhysOverridden)
             {
-                CFOverridden = false;
-                SAVControl.SetProgramVariable("OverrideConstantForce", (int)SAVControl.GetProgramVariable("OverrideConstantForce") - 1);
+                ApplyPhysOverridden = false;
+                if (SAVControl) { SAVControl.SetProgramVariable("DisablePhysicsApplication", (int)SAVControl.GetProgramVariable("DisablePhysicsApplication") - 1); }
             }
         }
         public void SFEXT_G_RespawnButton()
         {
-            if (CFOverridden)
+            if (ApplyPhysOverridden)
             {
-                CFOverridden = false;
-                SAVControl.SetProgramVariable("OverrideConstantForce", (int)SAVControl.GetProgramVariable("OverrideConstantForce") - 1);
+                ApplyPhysOverridden = false;
+                if (SAVControl) { SAVControl.SetProgramVariable("DisablePhysicsApplication", (int)SAVControl.GetProgramVariable("DisablePhysicsApplication") - 1); }
             }
         }
         public void SFEXT_O_TakeOwnership()
